+++ package.json
@@ -10,6 +10,8 @@
     "bootstrap": "^4.3.1",
     "cross-env": "^6.0.3",
     "font-awesome": "^4.7.0",
+    "moment": "^2.24.0",
+    "moment-timezone": "^0.5.27",
     "node-sass": "^4.13.0",
     "prop-types": "^15.7.2",
     "react": "^16.11.0",
